Utilities: Rewrite AppleEfiSignTool to use AppleVerificationLib

+++ Library/OcPeCoffLib/PeCoffRelocate.c
@@ -667,7 +667,7 @@ PeCoffRelocateImageForRuntime (
       "The following accesses must be performed unaligned."
       );
 
-    ASSERT (sizeof (EFI_IMAGE_BASE_RELOCATION_BLOCK) <= MAX_UINT32 - RelocWalker->SizeOfBlock);
+    ASSERT (sizeof (EFI_IMAGE_BASE_RELOCATION_BLOCK) <= (UINT32) (MAX_UINT32 - RelocWalker->SizeOfBlock));
 
     SizeOfRelocs = RelocWalker->SizeOfBlock - sizeof (EFI_IMAGE_BASE_RELOCATION_BLOCK);
